+++ app/src/main/java/app/passwordstore/ui/sshkeygen/SshKeyGenActivity.kt
@@ -118,6 +118,7 @@ class SshKeyGenActivity : AppCompatActivity() {
                 BiometricAuthenticator.authenticate(
                   this@SshKeyGenActivity,
                   R.string.biometric_prompt_title_ssh_keygen,
+                  allowPin = true,
                 ) { result ->
                   // Do not cancel on failed attempts as these are handled by the
                   // authenticator UI.